wedf

+++ uni-edu/lab1/Task1/task1.1/file1.go
@@ -3,7 +3,7 @@ package main
 import "fmt"
 
 func main() {
-	var arr = [5]int{1, 2, 3, 4, 5}
+	var arr = [5]float64{1, 2, 3, 4, 5}
 	fmt.Println(arr)
 }
 //output: [1 2 3 4 5]
